% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramGUI.R
\name{calcE}
\alias{calcE}
\title{Calculates a matrix in which each column is a skewed Gaussian}
\usage{
calcE(theta, lambda)
}
\arguments{
\item{theta}{vector of parameter estimates}

\item{lambda}{wavelengths at which to calculate model}
}
\value{
\code{matrix}
}
\description{
Calculates a matrix in which each column is a skewed Gaussian.
Like \code{calcEhiergaus} from TIMP package but uses a vector not a list of parameter estimates.
}
