% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\alias{datamat}
\alias{times}
\alias{waves}
\title{This is an example dataset included in this package}
\description{
Dispersion corrected time-resolved transient-absorption data
of the peridinin chlorophyll protein (PCP) excited with 490 nm laser light
from the publication of Stokkum et.al. (2009)
}
\references{
\doi{10.1016/j.chemphys.2008.10.005}
}
\author{
Ivo van Stokkum \email{i.h.m.van.stokkum@vu.nl}
}
\keyword{data}
