% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_compressed}
\alias{is_compressed}
\title{is_compressed}
\usage{
is_compressed(filename, magic.number = as.raw(c("0x1f", "0x8b")))
}
\arguments{
\item{filename}{The filename of the file to test for magic compression codes}

\item{magic.number}{The magic numbers in as a vector of strings with the hexidecimal numbers (e.g. "0x1f")}
}
\value{
boolean, TRUE if the file is compressed
}
\description{
Helper function for is_rdata, checks if the file is a compressed (gzip) file. Does not (yet) check for bzip2 or xz compression.
}
