% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_rdata}
\alias{is_rdata}
\title{is_rdata}
\usage{
is_rdata(filename)
}
\arguments{
\item{filename}{The filename of the file to test if it is an rdata file}
}
\value{
boolean, TRUE if the file is an rdata file
}
\description{
Checks a file is a rdata file by inspecting the file for so called magic bytes
}

