% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramGUI.R
\name{kroneckercol}
\alias{kroneckercol}
\title{kroneckercol: column-wise kronecker product}
\usage{
kroneckercol(A, B)
}
\arguments{
\item{A}{numerical matrix}

\item{B}{numerical matrix}
}
\value{
column-wise kronecker product of A and B
}
\description{
The column-wise kronecker product is also called the Khatri–Rao product
}

