% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramGUI.R
\name{plotterforGUI}
\alias{plotterforGUI}
\title{Master plot function for paramGUI}
\usage{
plotterforGUI(
  modtype = "kin",
  X = matrix(),
  data,
  model,
  theta = vector(),
  result,
  lin = NA,
  mu = 0,
  guessIRF = FALSE
)
}
\arguments{
\item{modtype}{either 'kin', 'spec' or 'spectemp'}

\item{X}{matrix of conditionally linear parameters, if any}

\item{data}{object of class \code{dat} containing data}

\item{model}{object of class \code{dat} containing data}

\item{theta}{object of class \code{theta} containing parameters}

\item{result}{object returned by \code{fitModel} or in the case
\code{modtype=='spectemp'}, by \code{nls}}

\item{lin}{The linear range for the concentration plot}

\item{mu}{The center of the lin-log axis is \code{lin} is specified}

\item{guessIRF}{Boolean to indicate whether to try and guess the location of the IRF}
}
\value{
graphics
}
\description{
Master plot function for paramGUI
}
