% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_dataset}
\alias{datamat}
\alias{example_dataset}
\alias{times}
\alias{waves}
\title{This is data to be included in my package}
\description{
Dispersion corrected time-resolved transient-absoprtion data 
of the peridinin chlorophyll protein (PCP) excited with 490 nm laser light 
from the publication of Stokkum et.al. (2009)
}
\author{
Ivo van Stokkum \email{i.h.m.van.stokkum@vu.nl}
}
\references{
\url{https://dx.doi.org/10.1016/j.chemphys.2008.10.005}
}
\keyword{data}

